+++ src/main/java/com/deepnyangning/capstonebe/domain/access/entity/AccessLog.java
@@ -29,7 +29,7 @@ public class AccessLog {
 
     private LocalDateTime accessTime;
 
-    private float similarity;
+    private double similarity;
 
     @PrePersist
     protected void onCreate(){
