+++ src/main/java/com/deepnyangning/capstonebe/domain/access/entity/FailLog.java
@@ -22,7 +22,7 @@ public class FailLog {
 
     private LocalDateTime accessTime;
 
-    private Float similarity;
+    private Double similarity;
 
     @PrePersist
     protected void onCreate(){
